Escalonamento modificado pra ser igual ao apresentado. O campo PC do PCB foi modificado para timeNecessary, que representa o valor do quantum necessário pra finalizar a tarefa.

+++ src/fileHandler/FileHandler.java
@@ -28,7 +28,7 @@ public class FileHandler {
 
             while (reader.hasNext()){
                 PCB pcb = new PCB(Integer.parseInt(reader.nextLine()),reader.nextLine(),
-                        Integer.parseInt(reader.nextLine()),Integer.parseInt(reader.nextLine()));
+                        Float.parseFloat(reader.nextLine()),Integer.parseInt(reader.nextLine()));
                 data.list.add(pcb);
 
             }
